+++ build_modules/lib/src/kernel_builder.dart
@@ -16,6 +16,7 @@ import 'common.dart';
 import 'errors.dart';
 import 'module_builder.dart';
 import 'modules.dart';
+import 'platform.dart';
 import 'scratch_space.dart';
 import 'workers.dart';
 
@@ -31,6 +32,8 @@ class KernelBuilder implements Builder {
 
   final String outputExtension;
 
+  final DartPlatform platform;
+
   /// Whether this should create summary kernel files or full kernel files.
   ///
   /// Summary files only contain the "outline" of the module - you can think of
@@ -41,11 +44,12 @@ class KernelBuilder implements Builder {
   final String sdkKernelPath;
 
   KernelBuilder(
-      {@required this.summaryOnly,
+      {@required this.platform,
+      @required this.summaryOnly,
       @required this.sdkKernelPath,
       @required this.outputExtension})
       : buildExtensions = {
-          moduleExtension: [outputExtension]
+          moduleExtension(platform): [outputExtension]
         };
 
   @override