+++ build_modules/lib/src/modules.g.dart
@@ -11,6 +11,7 @@ Module _$ModuleFromJson(Map<String, dynamic> json) {
       _assetIdFromJson(json['p'] as List),
       _assetIdsFromJson(json['s'] as List),
       _assetIdsFromJson(json['d'] as List),
+      _platformFromJson(json['pf'] as String),
       isMissing: json['m'] as bool ?? false);
 }
 
@@ -18,5 +19,6 @@ Map<String, dynamic> _$ModuleToJson(Module instance) => <String, dynamic>{
       'p': _assetIdToJson(instance.primarySource),
       's': _assetIdsToJson(instance.sources),
       'd': _assetIdsToJson(instance.directDependencies),
-      'm': instance.isMissing
+      'm': instance.isMissing,
+      'pf': _platformToJson(instance.platform)
     };