+++ build_modules/test/meta_module_builder_test.dart
@@ -8,18 +8,22 @@ import 'package:test/test.dart';
 
 import 'package:build_modules/build_modules.dart';
 import 'package:build_modules/src/meta_module.dart';
+import 'package:build_modules/src/module_library.dart';
+import 'package:build_modules/src/platform.dart';
 
 import 'matchers.dart';
 
 main() {
   test('can serialize meta modules', () async {
     var assetA = AssetId('a', 'lib/a.dart');
-    var moduleA = Module(assetA, [assetA], <AssetId>[]);
+    var moduleA = Module(assetA, [assetA], <AssetId>[], DartPlatform.dart2js);
     var metaA = MetaModule([moduleA]);
-    await testBuilder(MetaModuleBuilder(), {
-      'a|lib/a.module.library': 'true\n\n',
+    await testBuilder(MetaModuleBuilder(DartPlatform.dart2js), {
+      'a|lib/a.module.library':
+          ModuleLibrary.fromSource(assetA, '').serialize(),
     }, outputs: {
-      'a|lib/$metaModuleExtension': encodedMatchesMetaModule(metaA),
+      'a|lib/${metaModuleExtension(DartPlatform.dart2js)}':
+          encodedMatchesMetaModule(metaA),
     });
   });
 }