+++ build_modules/test/meta_module_clean_builder_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:convert';
+
 import 'package:build/build.dart';
 import 'package:build_modules/src/meta_module_clean_builder.dart';
 import 'package:build_test/build_test.dart';
@@ -11,63 +12,69 @@ import 'package:test/test.dart';
 
 import 'package:build_modules/build_modules.dart';
 import 'package:build_modules/src/meta_module.dart';
+import 'package:build_modules/src/platform.dart';
 
 import 'matchers.dart';
 
 main() {
   final assetA = AssetId('a', 'lib/a.dart');
   final assetB = AssetId('b', 'lib/b.dart');
+  final platform = DartPlatform.dart2js;
 
   test('unconnected components stay disjoint', () async {
-    var moduleA = Module(assetA, [assetA], []);
-    var moduleB = Module(assetB, [assetB], []);
+    var moduleA = Module(assetA, [assetA], [], platform);
+    var moduleB = Module(assetB, [assetB], [], platform);
 
     var metaA = MetaModule([moduleA]);
     var metaB = MetaModule([moduleB]);
 
-    await testBuilder(MetaModuleCleanBuilder(), {
-      'a|lib/$metaModuleExtension': json.encode(metaA),
-      'b|lib/$metaModuleExtension': json.encode(metaB),
+    await testBuilder(MetaModuleCleanBuilder(platform), {
+      'a|lib/${metaModuleExtension(DartPlatform.dart2js)}': json.encode(metaA),
+      'b|lib/${metaModuleExtension(platform)}': json.encode(metaB),
       'a|lib/a.dart': 'import "package:b/b.dart"',
       'b|lib/b.dart': 'import "package:a/a.dart"',
     }, outputs: {
-      'a|lib/$metaModuleCleanExtension': encodedMatchesMetaModule(metaA),
-      'b|lib/$metaModuleCleanExtension': encodedMatchesMetaModule(metaB),
+      'a|lib/${metaModuleCleanExtension(platform)}':
+          encodedMatchesMetaModule(metaA),
+      'b|lib/${metaModuleCleanExtension(platform)}':
+          encodedMatchesMetaModule(metaB),
     });
   });
 
   test('can handle cycles', () async {
-    var moduleA = Module(assetA, [assetA], [assetB]);
-    var moduleB = Module(assetB, [assetB], [assetA]);
+    var moduleA = Module(assetA, [assetA], [assetB], platform);
+    var moduleB = Module(assetB, [assetB], [assetA], platform);
 
     var metaA = MetaModule([moduleA]);
     var metaB = MetaModule([moduleB]);
     var clean = MetaModule([
-      Module(assetA, [assetA, assetB], [])
+      Module(assetA, [assetA, assetB], [], platform)
     ]);
 
-    await testBuilder(MetaModuleCleanBuilder(), {
-      'a|lib/$metaModuleExtension': json.encode(metaA),
-      'b|lib/$metaModuleExtension': json.encode(metaB),
+    await testBuilder(MetaModuleCleanBuilder(platform), {
+      'a|lib/${metaModuleExtension(platform)}': json.encode(metaA),
+      'b|lib/${metaModuleExtension(platform)}': json.encode(metaB),
       'a|lib/a.dart': 'import "package:b/b.dart"',
       'b|lib/b.dart': 'import "package:a/a.dart"',
     }, outputs: {
-      'a|lib/$metaModuleCleanExtension': encodedMatchesMetaModule(clean),
-      'b|lib/$metaModuleCleanExtension':
+      'a|lib/${metaModuleCleanExtension(platform)}':
+          encodedMatchesMetaModule(clean),
+      'b|lib/${metaModuleCleanExtension(platform)}':
           encodedMatchesMetaModule(MetaModule([])),
     });
   });
 
   test('Warns about missing .meta_module.raw files from dependencies',
       () async {
-    var moduleA = Module(assetA, [assetA], [assetB]);
+    var moduleA = Module(assetA, [assetA], [assetB], platform);
     var metaA = MetaModule([moduleA]);
     var logs = <LogRecord>[];
-    await testBuilder(MetaModuleCleanBuilder(), {
-      'a|lib/$metaModuleExtension': json.encode(metaA),
+    await testBuilder(MetaModuleCleanBuilder(platform), {
+      'a|lib/${metaModuleExtension(platform)}': json.encode(metaA),
       'a|lib/a.dart': 'import "package:b/b.dart"',
     }, outputs: {
-      'a|lib/$metaModuleCleanExtension': encodedMatchesMetaModule(metaA),
+      'a|lib/${metaModuleCleanExtension(platform)}':
+          encodedMatchesMetaModule(metaA),
     }, onLog: (r) {
       if (r.level >= Level.WARNING) logs.add(r);
     });