+++ build_modules/test/meta_module_test.dart
@@ -13,11 +13,13 @@ import 'package:build_modules/src/common.dart';
 import 'package:build_modules/src/meta_module.dart';
 import 'package:build_modules/src/module_library.dart';
 import 'package:build_modules/src/modules.dart';
+import 'package:build_modules/src/platform.dart';
 
 import 'matchers.dart';
 
 void main() {
   InMemoryAssetReader reader;
+  final defaultPlatform = DartPlatform.dart2js;
 
   List<AssetId> makeAssets(Map<String, String> assetDescriptors) {
     reader = InMemoryAssetReader();
@@ -31,20 +33,26 @@ void main() {
   }
 
   Future<MetaModule> metaModuleFromSources(
-      InMemoryAssetReader reader, List<AssetId> sources) async {
-    final libraries = (await Future.wait(sources.map((s) async =>
-            ModuleLibrary.fromSource(s, await reader.readAsString(s)))))
+      InMemoryAssetReader reader, List<AssetId> sources,
+      {DartPlatform platform}) async {
+    platform ??= defaultPlatform;
+    final libraries = (await Future.wait(sources
+            .where((s) => s.package != r'$sdk')
+            .map((s) async =>
+                ModuleLibrary.fromSource(s, await reader.readAsString(s)))))
         .where((l) => l.isImportable);
     for (final library in libraries) {
       reader.cacheStringAsset(
-          library.id.changeExtension(moduleLibraryExtension), '$library');
+          library.id.changeExtension(moduleLibraryExtension),
+          '${library.serialize()}');
     }
     return MetaModule.forLibraries(
         reader,
         libraries
             .map((l) => l.id.changeExtension(moduleLibraryExtension))
             .toList(),
-        ModuleStrategy.coarse);
+        ModuleStrategy.coarse,
+        platform);
   }
 
   test('no strongly connected components, one shared lib', () async {
@@ -68,9 +76,9 @@ void main() {
     var d = AssetId('myapp', 'lib/src/d.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a], [b, c])),
-      matchesModule(Module(b, [b], [c])),
-      matchesModule(Module(c, [c, d], [])),
+      matchesModule(Module(a, [a], [b, c], defaultPlatform)),
+      matchesModule(Module(b, [b], [c], defaultPlatform)),
+      matchesModule(Module(c, [c, d], [], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -96,7 +104,7 @@ void main() {
     var c = AssetId('myapp', 'lib/src/c.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a, b, c], []))
+      matchesModule(Module(a, [a, b, c], [], defaultPlatform))
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -140,9 +148,9 @@ void main() {
     var f = AssetId('myapp', 'lib/src/f.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a, c], [g, e])),
-      matchesModule(Module(b, [b, d], [c, e, g])),
-      matchesModule(Module(e, [e, g, f], [])),
+      matchesModule(Module(a, [a, c], [g, e], defaultPlatform)),
+      matchesModule(Module(b, [b, d], [c, e, g], defaultPlatform)),
+      matchesModule(Module(e, [e, g, f], [], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -160,11 +168,7 @@ void main() {
     var b = AssetId('b', 'lib/b.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [
-        a,
-      ], [
-        b
-      ])),
+      matchesModule(Module(a, [a], [b], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -192,8 +196,8 @@ void main() {
     var d = AssetId('myapp', 'lib/src/d.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a, c, d], [b])),
-      matchesModule(Module(b, [b], [])),
+      matchesModule(Module(a, [a, c, d], [b], defaultPlatform)),
+      matchesModule(Module(b, [b], [], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -233,12 +237,12 @@ void main() {
     var f = AssetId('myapp', 'lib/src/f.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a], [d, e, f])),
-      matchesModule(Module(b, [b], [d, e])),
-      matchesModule(Module(c, [c], [d, f])),
-      matchesModule(Module(d, [d], [])),
-      matchesModule(Module(e, [e], [d])),
-      matchesModule(Module(f, [f], [d])),
+      matchesModule(Module(a, [a], [d, e, f], defaultPlatform)),
+      matchesModule(Module(b, [b], [d, e], defaultPlatform)),
+      matchesModule(Module(c, [c], [d, f], defaultPlatform)),
+      matchesModule(Module(d, [d], [], defaultPlatform)),
+      matchesModule(Module(e, [e], [d], defaultPlatform)),
+      matchesModule(Module(f, [f], [d], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -266,7 +270,7 @@ void main() {
     var sap = AssetId('myapp', 'lib/src/a.part.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a, ap, sap], [])),
+      matchesModule(Module(a, [a, ap, sap], [], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -294,9 +298,9 @@ void main() {
     var c = AssetId('myapp', 'lib/src/c.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a], [])),
-      matchesModule(Module(b, [b, c], [])),
-      matchesModule(Module(sa, [sa], [c])),
+      matchesModule(Module(a, [a], [], defaultPlatform)),
+      matchesModule(Module(b, [b, c], [], defaultPlatform)),
+      matchesModule(Module(sa, [sa], [c], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -328,9 +332,9 @@ void main() {
     var d = AssetId('myapp', 'web/d.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a], [b, c])),
-      matchesModule(Module(b, [b], [c])),
-      matchesModule(Module(c, [c, d], [])),
+      matchesModule(Module(a, [a], [b, c], defaultPlatform)),
+      matchesModule(Module(b, [b], [c], defaultPlatform)),
+      matchesModule(Module(c, [c, d], [], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
@@ -369,48 +373,81 @@ void main() {
     var e = AssetId('myapp', 'web/e.dart');
 
     var expectedModules = [
-      matchesModule(Module(a, [a, b], [c])),
-      matchesModule(Module(c, [c, d], [])),
-      matchesModule(Module(e, [e], [d])),
+      matchesModule(Module(a, [a, b], [c], defaultPlatform)),
+      matchesModule(Module(c, [c, d], [], defaultPlatform)),
+      matchesModule(Module(e, [e], [d], defaultPlatform)),
     ];
 
     var meta = await metaModuleFromSources(reader, assets);
     expect(meta.modules, unorderedMatches(expectedModules));
   });
 
-  test('conditional import directive added to module dependencies', () async {
+  test(
+      'conditional import directives are added to module dependencies based '
+      'on platform', () async {
     var assets = makeAssets({
       'myapp|web/a.dart': '''
-        import 'b.dart'
-        if (expression1) 'b1.dart'
-        if (expression2) 'b2.dart'
-        if (expression3) 'b3.dart';
+        import 'default.dart'
+        if (dart.library.ui) 'ui.dart'
+        if (dart.library.io) 'io.dart'
+        if (dart.library.html) 'html.dart';
       ''',
-      'myapp|web/b.dart': '',
-      'myapp|web/b1.dart': '''
+      'myapp|web/default.dart': '',
+      'myapp|web/io.dart': '''
       ''',
-      'myapp|web/b2.dart': '''
+      'myapp|web/html.dart': '''
       ''',
-      'myapp|web/b3.dart': '''
+      'myapp|web/ui.dart': '''
       ''',
     });
 
-    var a = AssetId('myapp', 'web/a.dart');
-    var b = AssetId('myapp', 'web/b.dart');
-    var b1 = AssetId('myapp', 'web/b1.dart');
-    var b2 = AssetId('myapp', 'web/b2.dart');
-    var b3 = AssetId('myapp', 'web/b3.dart');
+    var primaryId = AssetId('myapp', 'web/a.dart');
+    var defaultId = AssetId('myapp', 'web/default.dart');
+    var htmlId = AssetId('myapp', 'web/html.dart');
+    var ioId = AssetId('myapp', 'web/io.dart');
+    var uiId = AssetId('myapp', 'web/ui.dart');
 
-    var expectedModules = [
-      matchesModule(Module(a, [a], [b, b1, b2, b3])),
-      matchesModule(Module(b, [b], [])),
-      matchesModule(Module(b1, [b1], [])),
-      matchesModule(Module(b2, [b2], [])),
-      matchesModule(Module(b3, [b3], [])),
-    ];
+    var expectedModulesForPlatform = {
+      DartPlatform.dart2js: [
+        matchesModule(
+            Module(primaryId, [primaryId], [htmlId], DartPlatform.dart2js)),
+        matchesModule(Module(htmlId, [htmlId], [], DartPlatform.dart2js)),
+        matchesModule(Module(ioId, [ioId], [], DartPlatform.dart2js)),
+        matchesModule(Module(defaultId, [defaultId], [], DartPlatform.dart2js)),
+        matchesModule(Module(uiId, [uiId], [], DartPlatform.dart2js)),
+      ],
+      DartPlatform.dartdevc: [
+        matchesModule(
+            Module(primaryId, [primaryId], [htmlId], DartPlatform.dartdevc)),
+        matchesModule(Module(htmlId, [htmlId], [], DartPlatform.dartdevc)),
+        matchesModule(Module(ioId, [ioId], [], DartPlatform.dartdevc)),
+        matchesModule(
+            Module(defaultId, [defaultId], [], DartPlatform.dartdevc)),
+        matchesModule(Module(uiId, [uiId], [], DartPlatform.dartdevc)),
+      ],
+      DartPlatform.flutter: [
+        matchesModule(
+            Module(primaryId, [primaryId], [uiId], DartPlatform.flutter)),
+        matchesModule(Module(htmlId, [htmlId], [], DartPlatform.flutter)),
+        matchesModule(Module(ioId, [ioId], [], DartPlatform.flutter)),
+        matchesModule(Module(defaultId, [defaultId], [], DartPlatform.flutter)),
+        matchesModule(Module(uiId, [uiId], [], DartPlatform.flutter)),
+      ],
+      DartPlatform.vm: [
+        matchesModule(Module(primaryId, [primaryId], [ioId], DartPlatform.vm)),
+        matchesModule(Module(htmlId, [htmlId], [], DartPlatform.vm)),
+        matchesModule(Module(ioId, [ioId], [], DartPlatform.vm)),
+        matchesModule(Module(defaultId, [defaultId], [], DartPlatform.vm)),
+        matchesModule(Module(uiId, [uiId], [], DartPlatform.vm)),
+      ]
+    };
 
-    var meta = await metaModuleFromSources(reader, assets);
-
-    expect(meta.modules, unorderedMatches(expectedModules));
+    for (var platform in expectedModulesForPlatform.keys) {
+      var meta =
+          await metaModuleFromSources(reader, assets, platform: platform);
+      expect(
+          meta.modules, unorderedMatches(expectedModulesForPlatform[platform]),
+          reason: meta.modules.map((m) => m.toJson()).toString());
+    }
   });
 }