+++ build_modules/test/module_builder_test.dart
@@ -14,27 +14,30 @@ import 'package:build_modules/src/modules.dart';
 import 'matchers.dart';
 
 main() {
+  final platform = DartPlatform.dart2js;
+
   test('can serialize modules and only output for primary sources', () async {
     var assetA = AssetId('a', 'lib/a.dart');
     var assetB = AssetId('a', 'lib/b.dart');
     var assetC = AssetId('a', 'lib/c.dart');
     var assetD = AssetId('a', 'lib/d.dart');
     var assetE = AssetId('a', 'lib/e.dart');
-    var moduleA = Module(assetA, [assetA], <AssetId>[]);
-    var moduleB = Module(assetB, [assetB, assetC], <AssetId>[]);
-    var moduleD = Module(assetD, [assetD, assetE], <AssetId>[]);
+    var moduleA = Module(assetA, [assetA], <AssetId>[], platform);
+    var moduleB = Module(assetB, [assetB, assetC], <AssetId>[], platform);
+    var moduleD = Module(assetD, [assetD, assetE], <AssetId>[], platform);
     var metaModule = MetaModule([moduleA, moduleB, moduleD]);
-    await testBuilder(ModuleBuilder(), {
+    await testBuilder(ModuleBuilder(platform), {
       'a|lib/a.dart': '',
       'a|lib/b.dart': '',
       'a|lib/c.dart': '',
       'a|lib/d.dart': '',
       'a|lib/e.dart': '',
-      'a|lib/.meta_module.clean': jsonEncode(metaModule.toJson()),
+      'a|lib/${metaModuleCleanExtension(platform)}':
+          jsonEncode(metaModule.toJson()),
     }, outputs: {
-      'a|lib/a.module': encodedMatchesModule(moduleA),
-      'a|lib/b.module': encodedMatchesModule(moduleB),
-      'a|lib/d.module': encodedMatchesModule(moduleD),
+      'a|lib/a${moduleExtension(platform)}': encodedMatchesModule(moduleA),
+      'a|lib/b${moduleExtension(platform)}': encodedMatchesModule(moduleB),
+      'a|lib/d${moduleExtension(platform)}': encodedMatchesModule(moduleD),
     });
   });
 }