+++ build_modules/test/summary_builder_test.dart
@@ -2,16 +2,18 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:build_modules/build_modules.dart';
 import 'package:build_test/build_test.dart';
 import 'package:logging/logging.dart';
 import 'package:test/test.dart';
+
+import 'package:build_modules/build_modules.dart';
 
 import 'matchers.dart';
 import 'util.dart';
 
 main() {
   Map<String, dynamic> assets;
+  final platform = DartPlatform.dart2js;
 
   group('basic project', () {
     setUp(() async {
@@ -32,37 +34,42 @@ main() {
 
       // Set up all the other required inputs for this test.
       await testBuilderAndCollectAssets(const ModuleLibraryBuilder(), assets);
-      await testBuilderAndCollectAssets(MetaModuleBuilder(), assets);
-      await testBuilderAndCollectAssets(MetaModuleCleanBuilder(), assets);
-      await testBuilderAndCollectAssets(ModuleBuilder(), assets);
+      await testBuilderAndCollectAssets(MetaModuleBuilder(platform), assets);
+      await testBuilderAndCollectAssets(
+          MetaModuleCleanBuilder(platform), assets);
+      await testBuilderAndCollectAssets(ModuleBuilder(platform), assets);
     });
 
     test('can output unlinked analyzer summaries for modules under lib and web',
         () async {
       var expectedOutputs = <String, Matcher>{
-        'b|lib/b.unlinked.sum': HasUnlinkedUris(['package:b/b.dart']),
-        'a|lib/a.unlinked.sum': HasUnlinkedUris(['package:a/a.dart']),
-        'a|web/index.unlinked.sum':
+        'b|lib/b${unlinkedSummaryExtension(platform)}':
+            HasUnlinkedUris(['package:b/b.dart']),
+        'a|lib/a${unlinkedSummaryExtension(platform)}':
+            HasUnlinkedUris(['package:a/a.dart']),
+        'a|web/index${unlinkedSummaryExtension(platform)}':
             HasUnlinkedUris([endsWith('web/index.dart')]),
       };
-      await testBuilder(UnlinkedSummaryBuilder(), assets,
+      await testBuilder(UnlinkedSummaryBuilder(platform), assets,
           outputs: expectedOutputs);
     });
 
     test('can output linked analyzer summaries for modules under lib and web',
         () async {
       // Build the unlinked summaries first.
-      await testBuilderAndCollectAssets(UnlinkedSummaryBuilder(), assets);
+      await testBuilderAndCollectAssets(
+          UnlinkedSummaryBuilder(platform), assets);
 
       // Actual test for LinkedSummaryBuilder;
       var expectedOutputs = <String, Matcher>{
-        'b|lib/b.linked.sum': allOf(HasLinkedUris(['package:b/b.dart']),
+        'b|lib/b${linkedSummaryExtension(platform)}': allOf(
+            HasLinkedUris(['package:b/b.dart']),
             HasUnlinkedUris(['package:b/b.dart'])),
-        'a|lib/a.linked.sum': allOf(
+        'a|lib/a${linkedSummaryExtension(platform)}': allOf(
             HasLinkedUris(
                 unorderedEquals(['package:b/b.dart', 'package:a/a.dart'])),
             HasUnlinkedUris(['package:a/a.dart'])),
-        'a|web/index.linked.sum': allOf(
+        'a|web/index${linkedSummaryExtension(platform)}': allOf(
             HasLinkedUris(unorderedEquals([
               'package:b/b.dart',
               'package:a/a.dart',
@@ -70,7 +77,7 @@ main() {
             ])),
             HasUnlinkedUris([endsWith('web/index.dart')])),
       };
-      await testBuilder(LinkedSummaryBuilder(), assets,
+      await testBuilder(LinkedSummaryBuilder(platform), assets,
           outputs: expectedOutputs);
     });
   });
@@ -85,17 +92,19 @@ main() {
 
       // Set up all the other required inputs for this test.
       await testBuilderAndCollectAssets(const ModuleLibraryBuilder(), assets);
-      await testBuilderAndCollectAssets(MetaModuleBuilder(), assets);
-      await testBuilderAndCollectAssets(MetaModuleCleanBuilder(), assets);
-      await testBuilderAndCollectAssets(ModuleBuilder(), assets);
-      await testBuilderAndCollectAssets(UnlinkedSummaryBuilder(), assets);
+      await testBuilderAndCollectAssets(MetaModuleBuilder(platform), assets);
+      await testBuilderAndCollectAssets(
+          MetaModuleCleanBuilder(platform), assets);
+      await testBuilderAndCollectAssets(ModuleBuilder(platform), assets);
+      await testBuilderAndCollectAssets(
+          UnlinkedSummaryBuilder(platform), assets);
     });
 
     test('print an error if there are any missing transitive modules',
         () async {
       var expectedOutputs = <String, Matcher>{};
       var logs = <LogRecord>[];
-      await testBuilder(LinkedSummaryBuilder(), assets,
+      await testBuilder(LinkedSummaryBuilder(platform), assets,
           outputs: expectedOutputs, onLog: logs.add);
       expect(
           logs,