+++ build_web_compilers/test/dart2js_bootstrap_test.dart
@@ -12,6 +12,7 @@ import 'util.dart';
 
 main() {
   Map<String, dynamic> assets;
+  final platform = DartPlatform.dart2js;
 
   setUp(() async {
     assets = {
@@ -30,9 +31,9 @@ main() {
 
     // Set up all the other required inputs for this test.
     await testBuilderAndCollectAssets(const ModuleLibraryBuilder(), assets);
-    await testBuilderAndCollectAssets(MetaModuleBuilder(), assets);
-    await testBuilderAndCollectAssets(MetaModuleCleanBuilder(), assets);
-    await testBuilderAndCollectAssets(ModuleBuilder(), assets);
+    await testBuilderAndCollectAssets(MetaModuleBuilder(platform), assets);
+    await testBuilderAndCollectAssets(MetaModuleCleanBuilder(platform), assets);
+    await testBuilderAndCollectAssets(ModuleBuilder(platform), assets);
   });
 
   test('can bootstrap dart entrypoints', () async {