+++ build_web_compilers/test/dev_compiler_bootstrap_test.dart
@@ -12,6 +12,7 @@ import 'util.dart';
 
 main() {
   Map<String, dynamic> assets;
+  final platform = DartPlatform.dartdevc;
 
   setUp(() async {
     assets = {
@@ -31,11 +32,11 @@ main() {
 
     // Set up all the other required inputs for this test.
     await testBuilderAndCollectAssets(const ModuleLibraryBuilder(), assets);
-    await testBuilderAndCollectAssets(MetaModuleBuilder(), assets);
-    await testBuilderAndCollectAssets(MetaModuleCleanBuilder(), assets);
-    await testBuilderAndCollectAssets(ModuleBuilder(), assets);
-    await testBuilderAndCollectAssets(UnlinkedSummaryBuilder(), assets);
-    await testBuilderAndCollectAssets(LinkedSummaryBuilder(), assets);
+    await testBuilderAndCollectAssets(MetaModuleBuilder(platform), assets);
+    await testBuilderAndCollectAssets(MetaModuleCleanBuilder(platform), assets);
+    await testBuilderAndCollectAssets(ModuleBuilder(platform), assets);
+    await testBuilderAndCollectAssets(UnlinkedSummaryBuilder(platform), assets);
+    await testBuilderAndCollectAssets(LinkedSummaryBuilder(platform), assets);
     await testBuilderAndCollectAssets(DevCompilerBuilder(), assets);
   });