+++ build_web_compilers/test/dev_compiler_builder_test.dart
@@ -13,6 +13,7 @@ import 'util.dart';
 
 main() {
   Map<String, dynamic> assets;
+  final platform = DartPlatform.dartdevc;
 
   group('error free project', () {
     setUp(() async {
@@ -33,11 +34,13 @@ main() {
 
       // Set up all the other required inputs for this test.
       await testBuilderAndCollectAssets(const ModuleLibraryBuilder(), assets);
-      await testBuilderAndCollectAssets(MetaModuleBuilder(), assets);
-      await testBuilderAndCollectAssets(MetaModuleCleanBuilder(), assets);
-      await testBuilderAndCollectAssets(ModuleBuilder(), assets);
-      await testBuilderAndCollectAssets(UnlinkedSummaryBuilder(), assets);
-      await testBuilderAndCollectAssets(LinkedSummaryBuilder(), assets);
+      await testBuilderAndCollectAssets(MetaModuleBuilder(platform), assets);
+      await testBuilderAndCollectAssets(
+          MetaModuleCleanBuilder(platform), assets);
+      await testBuilderAndCollectAssets(ModuleBuilder(platform), assets);
+      await testBuilderAndCollectAssets(
+          UnlinkedSummaryBuilder(platform), assets);
+      await testBuilderAndCollectAssets(LinkedSummaryBuilder(platform), assets);
     });
 
     test('can compile ddc modules under lib and web', () async {
@@ -64,11 +67,14 @@ main() {
 
         // Set up all the other required inputs for this test.
         await testBuilderAndCollectAssets(const ModuleLibraryBuilder(), assets);
-        await testBuilderAndCollectAssets(MetaModuleBuilder(), assets);
-        await testBuilderAndCollectAssets(MetaModuleCleanBuilder(), assets);
-        await testBuilderAndCollectAssets(ModuleBuilder(), assets);
-        await testBuilderAndCollectAssets(UnlinkedSummaryBuilder(), assets);
-        await testBuilderAndCollectAssets(LinkedSummaryBuilder(), assets);
+        await testBuilderAndCollectAssets(MetaModuleBuilder(platform), assets);
+        await testBuilderAndCollectAssets(
+            MetaModuleCleanBuilder(platform), assets);
+        await testBuilderAndCollectAssets(ModuleBuilder(platform), assets);
+        await testBuilderAndCollectAssets(
+            UnlinkedSummaryBuilder(platform), assets);
+        await testBuilderAndCollectAssets(
+            LinkedSummaryBuilder(platform), assets);
       });
 
       test('reports useful messages', () async {
@@ -98,9 +104,10 @@ main() {
 
         // Set up all the other required inputs for this test.
         await testBuilderAndCollectAssets(const ModuleLibraryBuilder(), assets);
-        await testBuilderAndCollectAssets(MetaModuleBuilder(), assets);
-        await testBuilderAndCollectAssets(MetaModuleCleanBuilder(), assets);
-        await testBuilderAndCollectAssets(ModuleBuilder(), assets);
+        await testBuilderAndCollectAssets(MetaModuleBuilder(platform), assets);
+        await testBuilderAndCollectAssets(
+            MetaModuleCleanBuilder(platform), assets);
+        await testBuilderAndCollectAssets(ModuleBuilder(platform), assets);
       });
 
       test('reports useful messages', () async {